- Made CFS++ openSUSE 11.0 respectively GCC 4.3.x compatible. - Backported ProgramOptions class from NACS (rev. 734) and removed CommandLine handler of Uwe Fabricius.

+++ source/OLAS/graph/graphmanagersbmmat.cc
@@ -378,7 +378,7 @@ namespace OLAS {
     //         constraints
     UInt aux;
     for ( UInt i = 1; i <= length1; i++ ) {
-      aux = std::abs( connect1[i] );
+      aux = std::abs( (double) connect1[i] );
       if ( aux > 0 ) {
         if ( aux <= numLastFreeDof_[pdeID1] ) {
           vertexList1_.push_back( aux );
@@ -393,7 +393,7 @@ namespace OLAS {
     //         the graph and one for the IDBCgraph (which handles the dofs
     //         fixed by inhomogeneous Dirichlet boundary conditions)
     for ( UInt i = 1; i <= length2; i++ ) {
-      aux = std::abs( connect2[i] );
+      aux = std::abs( (double) connect2[i] );
       if ( aux > 0 ) {
         if ( aux > numLastFreeDof_[pdeID2] ) {
           edgeList2_.push_back( aux - numLastFreeDof_[pdeID2] );
